+++ src/components/layout/index.js
@@ -1,6 +1,5 @@
 export * from '../BackToTopButton';
 export * from '../Footer';
-export * from '../Header';
 export * from '../HeroSection/HeroSection';
 export * from '../LandingPage';
 export * from '../MainMenu';